Added -i argument

+++ server/db/seeder.js
@@ -8,31 +8,47 @@ const MongooseConnection = require('./connection');
 
 require('dotenv').config({ path: path.resolve(__dirname, '../../config/config.env') });
 
-MongooseConnection.connect();
+if (process.env.LOCAL_DEVELOPMENT) {
+  MongooseConnection.connect();
 
-const importData = async () => {
-  try {
-    await User.deleteMany();
-    await WishCard.deleteMany();
-    await User.insertMany(users);
-    await WishCard.insertMany(wishcards);
-    process.exit();
-  } catch (error) {
-    process.exit(1);
-  }
-};
+  const insertData = async () => {
+    try {
+      await User.insertMany(users);
+      await WishCard.insertMany(wishcards);
+      process.exit();
+    } catch (error) {
+      process.exit(1);
+    }
+  };
 
-const destroyData = async () => {
-  try {
-    await User.deleteMany();
-    await WishCard.deleteMany();
-    process.exit();
-  } catch (error) {
-    process.exit(1);
+  const deleteDataAndImport = async () => {
+    try {
+      await User.deleteMany();
+      await WishCard.deleteMany();
+      await User.insertMany(users);
+      await WishCard.insertMany(wishcards);
+      process.exit();
+    } catch (error) {
+      process.exit(1);
+    }
+  };
+
+  const destroyData = async () => {
+    try {
+      await User.deleteMany();
+      await WishCard.deleteMany();
+      process.exit();
+    } catch (error) {
+      process.exit(1);
+    }
+  };
+  if (process.argv[2] === '-d') {
+    destroyData();
+  } else if (process.argv[2] === '-i') {
+    insertData();
+  } else {
+    deleteDataAndImport();
   }
-};
-if (process.argv[2] === '-d') {
-  destroyData();
 } else {
-  importData();
+  process.exit();
 }
